+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-06-24 18:49:17
-// Commit: fix(network): prevent double tap duplicate action in network
+// Updated at: 2026-06-24 23:18:35
+// Commit: feat(network): add bi-directional event bus handler for network
 
 class networkComponentState {
-  final String commitRef = '2274';
-  final String lastMessage = 'fix(network): prevent double tap duplicate action in network';
-  final DateTime timestamp = DateTime.parse('2026-06-24T18:49:17+05:00');
+  final String commitRef = '2283';
+  final String lastMessage = 'feat(network): add bi-directional event bus handler for network';
+  final DateTime timestamp = DateTime.parse('2026-06-24T23:18:35+05:00');
 }